backend/app/ai/advanced_pipeline/requirements.txt
torch
torchvision
torchaudio
timm
transformers
opencv-python
scipy
trimesh
flask
numpy
accelerate
bitsandbytes
safetensors
pillow
pyyaml
sentence-transformers
scikit-learn
